Fix map zone and room handling
